Refactored VerifyPhoneNumberCode
