+++ native/src/feature/onboarding/component/styles.ts
